create folders, rename files
