+++ src/components/tree/phyloTree/phyloTree.js
@@ -3,8 +3,8 @@ import _debounce from "lodash/debounce";
 import { event } from "d3-selection";
 import { min, max, sum } from "d3-array";
 import { scaleLinear } from "d3-scale";
-import { flattenTree, appendParentsToTree } from "./treeHelpers";
-import { dataFont, darkGrey } from "../../globalStyles";
+import { flattenTree, appendParentsToTree } from "../treeHelpers";
+import { dataFont, darkGrey } from "../../../globalStyles";
 
 /*
  * adds the total number of descendant leaves to each node in the tree
